+++ src/app/(routes)/admin/_utils/rdc-vision-helpers.ts
@@ -32,6 +32,7 @@ export const handleAnalyzeBtnClick = async (
     const visionResult = await analyzeScreenShot(
       base64FileContent,
       sessionPlayers,
+      3,
     );
 
     switch (visionResult.status) {
